+++ examples/with_cache.ts
@@ -19,6 +19,8 @@ await Bun.sleep(1500);
 
 const third_request = await testWithCache.get("https://www.example.com", { cache: true });
 
+await testWithCache.disconnectCache();
+
 console.log("Cache Tests", {
     firstRequest: first_request.cached,
     secondRequest: second_request.cached,